Fix scrollToElement when CSS is disabled
When CSS is disabled UI elements and HTML elements are not normally
associated with each other to save space/avoid creating many extra
containers.  That then makes scrollToElement unworkable; therefor
images and text should be associated with their respective elements
regardless of whether or not CSS is enabled: this does not create
any new containers etc.