Invert background/foreground colors on HTMLLinks
With the default implementation links get their colors from the HTML
and the user cannot readily see which link is active.  This commit
automatically inverts the link background and foreground.  This
doesn't really work so well being done through the theme because
the color of the link can come from the body tag etc and is thus
set mostly by the HTMLComponents parsing not through the Style